train: Fix segmentation of sample images
This is the reason why the sample images that were being produced
on seemingly high quality models were so bad - previously we were removing
all transformations including image normalization from the transformation
set, then both visualizing and segmenting the image. This effectively meant
that the model was getting information in the wrong channels, leading to
bad classifications due to average pooling getting tripped up (even though
it could get some of the contours right).

Instead, we need to ensure that we normalize the images before segmenting
them, but only "view" the unnnormalized image. This way we get much higher
quality segmentations that make much more sense.